+++ src/converters.ts
@@ -21,7 +21,7 @@
  */
 
 import { JsonConverter } from './jsonConverter';
-import { TemplateVars } from './jsonEditor/jsonEditorState';
+import { TemplateVars } from './jsonContext';
 
 /**
  * Converts the supplied unknown to JSON, rendering any property names
